third_party/blink/web_tests/external/wpt/css/css-cascade/scope-shadow-expected.txt

This is a testharness.js-based test.
FAIL @scope can match :host assert_equals: expected "1" but got "auto"
FAIL @scope can match :host(...) assert_equals: expected "1" but got "auto"
Harness: the test ran to completion.
